requirements.txt
Flask
Gunicorn
Pillow
Ultralytics